# chore(lsp): Add benchmark for performance on a large real-world repo