--- a/.gitmodules
+++ b/.gitmodules
@@ -11,4 +11,4 @@
 	url = https://github.com/denoland/node_test.git
 [submodule "cli/bench/testdata/lsp_benchdata"]
 	path = cli/bench/testdata/lsp_benchdata
-	url = https://github.com/nathanwhit/deno_lsp_benchdata
+	url = https://github.com/denoland/deno_lsp_benchdata.git