… cli/bench/testdata/deco_apps_requests.json